ArticleListViewをコンポーネント化
